fix: resolve react-hooks/preserve-manual-memoization errors in App.tsx
Move the OpenCode reconnect logic from an inline fetch in
handleToggleOpenCode (useCallback) into useOpenCodeModelSync as a
stable `reconnect` callback.  This lets the React Compiler preserve
the existing manual memoization — fixing the two errors that failed
CI run #1029.

Also removes the now-unnecessary eslint-disable directive and cleans
up unused imports (fetchOpenCodeModels, ModelOption) from App.tsx.
